moved DB creation out of imageCache into new schema,py
